+++ terraform/compute_images.tf
@@ -45,3 +45,44 @@ resource "google_compute_image" "example_source_image" {
   name = "example-source-image"
   source_image = data.google_compute_image.debian.self_link
 }
+
+resource "google_compute_instance" "testvm_snapshot" {
+  count = var.enable_testvm_not_accesible_internet_custom_image_snapshot ? 1 : 0
+
+  name         = "testvmusingsnapshot"
+  machine_type = "n1-standard-1"
+  zone         = "us-central1-a"
+  tags = ["testvmtag1", "testvmtag2"]
+
+  boot_disk {
+    initialize_params {
+      image = google_compute_image.example_snapshot_image.self_link
+    }
+  }
+
+  # boot_disk {
+  #   initialize_params {
+  #     image = "ubuntu-os-cloud/ubuntu-1804-lts"
+  #   }
+  # }
+
+  // Local SSD disk
+  scratch_disk {
+    interface = "SCSI"
+  }
+
+  network_interface {
+    network = "default"
+  }
+
+  metadata = {
+    foo = "bar"
+  }
+
+  metadata_startup_script = "echo Hello"
+
+  service_account {
+    email = var.client_email
+    scopes = ["userinfo-email", "compute-ro", "storage-ro"]
+  }
+}